% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocal.r
\name{cgwas}
\alias{cgwas}
\title{packagename: combine GWAS}
\usage{
cgwas(cgwas_dir, cgwas_data_dir, trait_data_name, trait_name, trait_num,
  snp_column_index, assoc_column_index, paral_num, plotFlag = 1,
  heteroprop = 1, Ccp = 0.1, Manco = 0.01, Locisep = 250000)
}
\arguments{
\item{cgwas_dir}{blablabla}

\item{cgwas_data_dir}{blablabla}

\item{trait_data_name}{blablabla}

\item{trait_name}{blablabla}

\item{trait_num}{blablabla}

\item{snp_column_index}{blablabla}

\item{assoc_column_index}{blablabla}

\item{paral_num}{blablabla}

\item{plotFlag}{indicate whether output the manhattan plot or bar plot}

\item{heteroprop}{blablabla}

\item{Ccp}{blablabla}

\item{Manco}{blablabla}

\item{Locisep}{blablabla}
}
\description{
This package provides a toolkit for combine GWAS
}
\details{
Integrating multiple single trait GWAS results

\code{CGWAS} combine multiple single trait GWAS results which indicated by its
arguments to calculate integrate association result.

\code{CGWAS} combine multiple single trait GWAS results to calculate integrate
association result from a set of trait combinations for each SNP specifically.
The method increase power detecting SNP having effect (may be weak in GWAS
for one trait) on multiple traits and type I error was controlled through
a large scale simulation analyses. Significant CGWAS result contains both
single trait GWAS hits and multiple traits combined GWAS hits after multiple test.

\code{snpFilename} shoule be a white-space (space or tab) delimited file
with header line. This file desctribes the SNP information and the first three
columun must be Chromesome, Base-pair position, SNP id.

Each data indicated by \code{gwasFilename} should be a white-space (space or tab)
delimited file with header line. The first two columns are BETA, P-vale and
the order of each line should align with the order of the \code{snpFilename}
file SNPs.

If \code{traitName} is NA, it will be filled by \code{gwasFilename}.

If \code{Gt} is NA, genome-wide threshold will be set as max(5e-8, 0.05/Rsnpnum).

If \code{St} is NA, suggestive threshold will be set as max(1.6e-6, 1/Rsnpnum).

\code{Rsnpnum} is set as the total number of SNPs.

\code{paral} set the function running in parallel mode and limit the maximum
number of unsable CPUs if it is set as an integer greater than 1. Otherwise,
the function will run in serial mode.
}
\examples{

cgwas_dir <- getwd()
cgwas_data_dir <- system.file("extdata", package = "CGWAS")
trait_data_name <- c("phe1.assoc", "phe2.assoc", "phe3.assoc", "phe4.assoc", "phe5.assoc")
trait_name <- c("phe1", "phe2", "phe3", "phe4", "phe5")
trait_num <-c(10015, 10015, 10015, 10015, 10015)
snp_column_index <- c(1,2,3)
assoc_column_index <- c(4,5)
paral_num <- 1
plotFlag <- 1
cgwas(cgwas_dir, cgwas_data_dir, trait_data_name, trait_name, trait_num,
      snp_column_index, assoc_column_index, paral_num)

}
