% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgwasProtocal.R
\name{cgwas}
\alias{cgwas}
\title{combine GWAS (C-GWAS)}
\usage{
cgwas(
  gwasFilePath,
  assocColInx,
  outputPath,
  mafFilePath = NULL,
  threadN = ceiling(parallel::detectCores()/2),
  traitName = NULL,
  indSNPN = 1e+06,
  simulTime = NULL,
  ebicoPwrInc = 1.1,
  sampleNInc = 0.7
)
}
\arguments{
\item{gwasFilePath}{A string list of path to GWASs summary data files.}

\item{assocColInx}{A number list indicates column number of SNP index (CHR &
BP & SNP) and association result (BETA & P).}

\item{outputPath}{Result output directory.}

\item{mafFilePath}{Path to allele frequency file.}

\item{threadN}{The number of threads for parallel computing. The default value
is half of the maximum number of cores on the hardware.}

\item{traitName}{A list of trait names.}

\item{indSNPN}{Independent SNPs number.}

\item{simulTime}{Number of simulations for multiple testing correction. This
parameter greatly affects the operating efficiency.}

\item{ebicoPwrInc}{Minimum EbICo power increase ratio.}

\item{sampleNInc}{Equivalent sample size in EbICo increase ratio.}
}
\value{
Several .csv files and figures will be generated in \code{outputPath}.

\code{MutipleTestingCorrection.txt}.

\code{GWAShits.csv}.

\code{EbICohits.csv}.

\code{SWaThits.csv}.

\code{Summaryhits.csv}.

\code{Summary.csv}.
}
\description{
A whole solution for illustrating multi-trait effect of all SNPs with a set
of GWASs summary data.
}
\details{
C-GWAS begins with GWASs summary as inputs and outputs a single vector of
combined p-values testing if the null is deviated. For each SNP, the null is
the absence of any effect on all traits, and the alternative is that its
effect deviates from 0 for at least one trait.

To estimate the effect of each SNP more accurately, several data organization
standard should be noticed. (1) The effect size in each GWAS should be in the
same scale. (2) In one GWAS, the minimum samples size of each SNP should
be no less than 60% of the median.

C-GWAS integrates two different statistical methods to ensure the optimal
power under various and complex scenarios while keeping a stable study-wide
type-I error rate. The first method uses an iterative effect based inversed
covariance weighting (i-EbICoW), which appears the most powerful when the
assumption (all SNPs share the same variance–covariance matrix of effect
sizes across traits) tends to be satisfied or moderately violated. The second
method is a truncated Wald test (TWT) which is more powerful than i-EbICoW
when the assumption tends to be severely violated. For each SNP, TWT proposes
the best subset of phenotypes by applying the Wald test to all subsets
determined under a series of preset thresholds.

C-GWAS controls the study-wide type-I error rate in an empirical manner via
simulations and adjust the resultant p-values in such a way that they are
directly comparable with those from traditional GWAS of a single trait.
}
\examples{

gwasFileDir <- system.file("extdata", package = "CGWAS")
gwasFileName <- c("phe1.assoc", "phe2.assoc", "phe3.assoc",
                  "phe4.assoc", "phe5.assoc")
gwasFilePath <- file.path(gwasFileDir, gwasFileName)
assocColInx <- c(1, 2, 3, 4, 5)
outputPath <- getwd()
traitName <- c("phe1", "phe2", "phe3", "phe4", "phe5")
simulTime <- 200

cgwas(gwasFilePath, assocColInx, outputPath,
      traitName = traitName, simulTime = simulTime)

}
